fix test cases

+++ src/main/java/com/driver/services/impl/ParkingLotServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ParkingLotServiceImpl implements ParkingLotService {
@@ -27,13 +28,11 @@ public class ParkingLotServiceImpl implements ParkingLotService {
 
     @Override
     public Spot addSpot(int parkingLotId, Integer numberOfWheels, Integer pricePerHour) {
-        SpotType spotType;
-        if (numberOfWheels == 2)
+        SpotType spotType = SpotType.OTHERS;
+        if (numberOfWheels.equals(2))
             spotType = SpotType.TWO_WHEELER;
-        else if (numberOfWheels == 4)
+        else if (numberOfWheels.equals(4))
             spotType = SpotType.FOUR_WHEELER;
-        else
-            spotType = SpotType.OTHERS;
 
         Spot spot = new Spot(spotType, pricePerHour);
         ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
@@ -45,34 +44,40 @@ public class ParkingLotServiceImpl implements ParkingLotService {
 
     @Override
     public void deleteSpot(int spotId) {
-        Spot spot = spotRepository1.findById(spotId).get();
-        ParkingLot parkingLot = spot.getParkingLot();
-        List<Spot> spotList= parkingLot.getSpotList();
-        for (Spot s : spotList) {
-            if(s == spot)
-                spotRepository1.delete(s);
+        Spot spot = spotRepository1.findById(spotId).orElse(null);
+        if (!Objects.isNull(spot)) {
+            ParkingLot parkingLot = spot.getParkingLot();
+            List<Spot> spotList= parkingLot.getSpotList();
+            for (Spot s : spotList) {
+                if(s == spot)
+                    spotRepository1.delete(s);
 
+            }
+            spotRepository1.deleteById(spotId);
         }
-        spotRepository1.deleteById(spotId);
     }
 
     @Override
     public Spot updateSpot(int parkingLotId, int spotId, int pricePerHour) {
-        Spot spot = spotRepository1.findById(spotId).get();
-        spot.setParkingLot(parkingLotRepository1.findById(parkingLotId).get());
-        spot.setPricePerHour(pricePerHour);
-        spotRepository1.save(spot);
-        return spot;
-
+        Spot spot = spotRepository1.findById(spotId).orElse(null);
+        if (!Objects.isNull(spot)) {
+            spot.setParkingLot(parkingLotRepository1.findById(parkingLotId).get());
+            spot.setPricePerHour(pricePerHour);
+            spotRepository1.save(spot);
+            return spot;
+        }
+        return null;
     }
 
     @Override
     public void deleteParkingLot(int parkingLotId) {
-        ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
-        List<Spot> spotList= parkingLot.getSpotList();
-        for (Spot spot: spotList) {
-            spotRepository1.delete(spot);
+        ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).orElse(null);
+        if (!Objects.isNull(parkingLot)) {
+            List<Spot> spotList= parkingLot.getSpotList();
+            for (Spot spot: spotList) {
+                spotRepository1.delete(spot);
+            }
+            parkingLotRepository1.deleteById(parkingLotId);
         }
-        parkingLotRepository1.deleteById(parkingLotId);
     }
 }
